+++ sshd-common/src/test/java/org/apache/sshd/common/config/keys/loader/pem/PKCS8PEMResourceKeyPairParserTest.java
@@ -26,7 +26,6 @@ import java.security.GeneralSecurityException;
 import java.security.KeyPair;
 import java.security.KeyPairGenerator;
 import java.security.PrivateKey;
-import java.security.PublicKey;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
@@ -43,7 +42,6 @@ import org.apache.sshd.util.test.JUnitTestSupport;
 import org.apache.sshd.util.test.NoIoTestCase;
 import org.junit.Assume;
 import org.junit.FixMethodOrder;
-import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
@@ -81,6 +79,11 @@ public class PKCS8PEMResourceKeyPairParserTest extends JUnitTestSupport {
         }
         if (SecurityUtils.isECCSupported()) {
             for (ECCurves curve : ECCurves.VALUES) {
+                if (!curve.isSupported()) {
+                    outputDebugMessage("Skip unsupported curve=%s", curve);
+                    continue;
+                }
+
                 params.add(new Object[] { KeyUtils.EC_ALGORITHM, curve.getKeySize() });
             }
         }
@@ -88,7 +91,7 @@ public class PKCS8PEMResourceKeyPairParserTest extends JUnitTestSupport {
     }
 
     @Test // see SSHD-760
-    public void testPkcs8() throws IOException, GeneralSecurityException {
+    public void testLocallyGeneratedPkcs8() throws IOException, GeneralSecurityException {
         KeyPairGenerator generator = SecurityUtils.getKeyPairGenerator(algorithm);
         if (keySize > 0) {
             generator.initialize(keySize);
@@ -114,14 +117,18 @@ public class PKCS8PEMResourceKeyPairParserTest extends JUnitTestSupport {
         }
     }
 
-    // see https://gist.github.com/briansmith/2ee42439923d8e65a266994d0f70180b
-    // openssl genpkey -algorithm EC -pkeyopt ec_paramgen_curve:P-256 -pkeyopt ec_param_enc:named_curve -out pkcs8-ecdsa-256.pem
-    // openssl ecparam -genkey -name prime256v1 -noout -out pkcs8-ec-256.key
-    // openssl pkcs8 -topk8 -inform PEM -outform PEM -nocrypt -in pkcs8-ec-256.key -out pkcs8-ec-256.pem
-    // openssl genpkey -algorithm RSA -pkeyopt rsa_keygen_bits:1024 -out pkcs8-rsa-1024.pem
-    // openssl asn1parse -inform PEM -in ...file... -dump
+    /*
+     * See https://gist.github.com/briansmith/2ee42439923d8e65a266994d0f70180b
+     *
+     * openssl genpkey -algorithm RSA -pkeyopt rsa_keygen_bits:1024 -out pkcs8-rsa-1024.pem
+     * openssl genpkey -algorithm EC -pkeyopt ec_paramgen_curve:P-256 -pkeyopt ec_param_enc:named_curve -out pkcs8-ecdsa-256.pem
+     *
+     * openssl ecparam -genkey -name prime256v1 -noout -out pkcs8-ec-256.key
+     * openssl pkcs8 -topk8 -inform PEM -outform PEM -nocrypt -in pkcs8-ec-256.key -out pkcs8-ec-256.pem
+     *
+     * openssl asn1parse -inform PEM -in ...file... -dump
+     */
     @Test // see SSHD-989
-    @Ignore("WIP")
     public void testPKCS8FileParsing() throws Exception {
         String resourceKey = "pkcs8-" + algorithm.toLowerCase() + "-" + keySize + ".pem";
         URL url = getClass().getResource(resourceKey);
@@ -129,18 +136,7 @@ public class PKCS8PEMResourceKeyPairParserTest extends JUnitTestSupport {
 
         Collection<KeyPair> pairs = PKCS8PEMResourceKeyPairParser.INSTANCE.loadKeyPairs(null, url, null);
         assertEquals("Mismatched extract keys count", 1, GenericUtils.size(pairs));
-
-        assertSignatureMatch("Cannot sign with recovered key pair", GenericUtils.head(pairs));
-    }
-
-    private static void assertSignatureMatch(String message, KeyPair kp) throws GeneralSecurityException {
-        assertSignatureMatch(message, kp.getPrivate(), kp.getPublic());
-    }
-
-    private static void assertSignatureMatch(
-            String message, PrivateKey privateKey, PublicKey publicKey)
-            throws GeneralSecurityException {
-        // TODO
+        validateKeyPairSignable(algorithm + "/" + keySize, GenericUtils.head(pairs));
     }
 
     @Override